Extract utils and widgets to their own files
